pyproject.toml
[project]
name = "project-aigis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "authlib==1.5.2",
    "boto3-stubs==1.38.9",
    "botocore-stubs==1.38.9",
    "cachetools==5.5.2",
    "certifi>=2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "cryptography==44.0.3",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "google-adk==0.4.0",
    "google-api-core==2.25.0rc0",
    "google-api-python-client==2.169.0",
    "google-auth==2.40.0",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.91.0",
    "google-cloud-bigquery==3.31.0",
    "google-cloud-core==2.4.3",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-secret-manager==2.23.3",
    "google-cloud-speech==2.32.0",
    "google-cloud-storage==2.19.0",
    "google-cloud-trace==1.16.1",
    "google-crc32c==1.7.1",
    "google-genai==1.13.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.20.3",
    "greenlet==3.2.1",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "levenshtein==0.27.1",
    "litellm==1.68.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mcp==1.7.1",
    "mdurl==0.1.2",
    "multidict==6.4.3",
    "mypy-boto3-bedrock-runtime==1.38.4",
    "numpy==2.2.5",
    "openai==1.75.0",
    "opentelemetry-api==1.32.1",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.32.1",
    "opentelemetry-semantic-conventions==0.53b1",
    "opik==1.7.16",
    "packaging==25.0",
    "pluggy==1.5.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rapidfuzz==3.13.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.0.0",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "sentry-sdk==2.27.0",
    "shapely==2.1.0",
    "six==1.17.0",
    "slack-bolt==1.23.0",
    "slack-sdk==3.35.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "sse-starlette==2.3.4",
    "starlette==0.46.2",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "types-awscrt==0.26.1",
    "types-s3transfer==0.12.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzlocal==5.3.1",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uuid6==2024.7.10",
    "uvicorn==0.34.2",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.21.0",
]
